почистил, разделил экшены и типы
