+++ src/components/app/app.tsx
@@ -1,5 +1,4 @@
 import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
 import { useAppDispatch, useAppSelector } from "../../utils/hooks/useAppStore";
 import { Routes, Route, useLocation, useNavigate} from 'react-router-dom';
 
@@ -14,7 +13,6 @@ import { getIngredients } from "../../service/actions/app";
 import { NotFound404, Home, Profile, Feed, Login, Register, ResetPassword, ForgotPassword } from "../../page";
 import { ProfileInputList } from "../profile/input-list/input-list";
 import { checkUserAuth } from "../../service/actions/user-auth";
-import { TListOrders, TIngredient } from "../../types";
 
 import appStyles from "./app.module.css";
