+++ src/components/profile/orders/profile-orders.tsx
@@ -1,7 +1,6 @@
 import { useState, useEffect } from "react";
 import { ListOrder } from "../../feed/list-order/list-order"
 import { useCookie } from "../../../utils/useCookie"
-import { useDispatch, useSelector } from "react-redux"
 import { connect, disconnect } from "../../../service/actions/ws-action"
 import { useAppDispatch, useAppSelector } from "../../../utils/hooks/useAppStore";
 import { TListOrders } from "../../../types";
