+++ src/constatnts/ws.ts
@@ -1,0 +1,30 @@
+const CONNECTING: 'CONNECTING' = 'CONNECTING';
+const ONLINE: 'ONLINE' = 'ONLINE';
+const OFFLINE: 'OFFLINE' = 'OFFLINE';
+
+export const WebsocketStatus = {
+  CONNECTING: CONNECTING,
+  ONLINE: ONLINE,
+  OFFLINE: OFFLINE,
+};
+
+export const LIVE_CONNECT: 'LIVE_CONNECT' = 'LIVE_CONNECT'
+export const LIVE_DISCONNECT: 'LIVE_DISCONNECT' = 'LIVE_DISCONNECT'
+
+export const WS_CONNECTING: "WS_CONNECTING" = "WS_CONNECTING";
+export const WS_CONNECTION_OPEN: 'WS_CONNECTION_OPEN' = 'WS_CONNECTION_OPEN';
+export const WS_CONNECTION_CLOSED: 'WS_CONNECTION_CLOSED' = 'WS_CONNECTION_CLOSED';
+export const WS_CONNECTION_ERROR: 'WS_CONNECTION_ERROR' = 'WS_CONNECTION_ERROR';
+export const WS_GET_MESSAGE: 'WS_GET_MESSAGE' = 'WS_GET_MESSAGE';
+export const WS_SEND_MESSAGE: 'WS_SEND_MESSAGE' = 'WS_SEND_MESSAGE';
+
+export const wsAction = {
+    connect: LIVE_CONNECT , 
+    disconnect: LIVE_DISCONNECT,
+    connecting: WS_CONNECTING,
+    open: WS_CONNECTION_OPEN,
+    closed: WS_CONNECTION_CLOSED,
+    error: WS_CONNECTION_ERROR,
+    getMessage: WS_GET_MESSAGE,
+    sendMessage: WS_SEND_MESSAGE,
+}
