+++ src/service/actions/burger-constructor.ts
@@ -1,27 +1,8 @@
 import request from "../../utils/request";
 import { ingrOption } from "../../utils/fetch-option";
-import { CLEANING_INGREDIENT_LIST } from "./constructor";
+import { AWAIT_ORDER, MAKE_ORDER_SUCCESS, MAKE_ORDER_FAILED, CLEANING_INGREDIENT_LIST } from "../../constatnts/actions";
 
 import { AppDispatch } from "..";
-
-export const AWAIT_ORDER: "AWAIT_ORDER" = "AWAIT_ORDER";
-export const MAKE_ORDER_SUCCESS: "MAKE_ORDER_SUCCESS" = "MAKE_ORDER_SUCCESS";
-export const MAKE_ORDER_FAILED: "MAKE_ORDER_FAILED" = "MAKE_ORDER_FAILED";
-
-type TOrderAwait = {
-  readonly type: typeof AWAIT_ORDER;
-}
-type TOrderSuccess = {
-  readonly type: typeof MAKE_ORDER_SUCCESS;
-  readonly orderNumber: number;
-}
-type TOrderFailed = {
-  readonly type: typeof MAKE_ORDER_FAILED;
-  readonly textErr: string;
-}
-
-export type TBurgerAction = TOrderAwait | TOrderSuccess | TOrderFailed;
-
 
 export function makeOrderApi(value: string[] | number []) {
   return function (dispatch: AppDispatch) {
