+++ src/service/actions/constructor.ts
@@ -1,28 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
 import { TIngredient } from '../../types';
-
-export const ADD_INGREDIENT: "ADD_INGREDIENT" = "ADD_INGREDIENT";
-export const DELETE_INGREDIENT: "DELETE_INGREDIENT" = "DELETE_INGREDIENT";
-export const MOVE_INGREDIENT: "MOVE_INGREDIENT" = "MOVE_INGREDIENT";
-export const CLEANING_INGREDIENT_LIST: "CLEANING_INGREDIENT_LIST" = "CLEANING_INGREDIENT_LIST"
-
-type TAddIngredient = {
-  readonly type: typeof ADD_INGREDIENT,
-  readonly ingr: TIngredient & {uniqueId: string}
-}
-type TDeleteIngredient = {
-  readonly type: typeof DELETE_INGREDIENT,
-  readonly ingr: number
-}
-type TMoveIngredient = {
-  readonly type: typeof MOVE_INGREDIENT,
-  readonly hoverIndex: number,
-  readonly dragIndex: number,
-}
-type TCleaninIngredientList = {
-  readonly type: typeof CLEANING_INGREDIENT_LIST
-}
-export type TConsructorAction = TAddIngredient | TDeleteIngredient | TMoveIngredient | TCleaninIngredientList
+import { ADD_INGREDIENT, DELETE_INGREDIENT, MOVE_INGREDIENT, CLEANING_INGREDIENT_LIST } from '../../constatnts/actions';
 
 export const addIngredient = (ingr: TIngredient) => ({
   type: ADD_INGREDIENT,
