+++ src/service/reducers/order-number.ts
@@ -1,10 +1,10 @@
-import { TOrderNumberAction as TAction } from "../actions/order-number";
 import {
-  GET_ORDDER_DETAILS,
   AWAIT_ORDDER_DETAILS,
   ERROR_ORDDER_DETAILS,
+  GET_ORDDER_DETAILS,
   COMPLETED_ORDDER_DETAILS,
-} from "../actions/order-number";
+} from "../../constatnts/actions";
+import { TOrderNumberAction as TAction } from "../actions/types";
 import { TInitialStateOrderNumber as TReducer } from "./types";
 
 const initialState: TReducer = {
@@ -13,7 +13,10 @@ const initialState: TReducer = {
   error: "",
 };
 
-export const getOrderDetailsReducer = (state = initialState, action: TAction) => {
+export const getOrderDetailsReducer = (
+  state = initialState,
+  action: TAction
+) => {
   switch (action.type) {
     case AWAIT_ORDDER_DETAILS: {
       return {
