+++ src/service/reducers/types.ts
@@ -1,4 +1,5 @@
-import { TIngredient, TType, TOrders, TRegistration } from "../../types";
+import { TIngredient, TType, TOrders } from "../../types";
+// import { TIngredient, TType, TOrders, TRegistration } from "../../types";
 
 export type TInitialStateApp = {
   dataRequest: boolean;
