+++ src/service/reducers/user-auth.ts
@@ -1,36 +1,34 @@
 import {
-  USER_LOGOUT,
   USER_DATA,
+  USER_LOGOUT,
   LOADING_STATUS,
-  TUserAuthAction as TAction,
-} from "../actions/user-auth";
+} from "../../constatnts/actions";
+import { TUserAuthAction as TAction } from "../actions/types";
 import { TInitialStateUserAuth as TReducer } from "./types";
 
 const initialState: TReducer = {
-  userData: {email: '', name: ''},
-  loading: {status: false, message: ''},
+  userData: { email: "", name: "" },
+  loading: { status: false, message: "" },
   // isLoggedIn: false,
 };
 
 export const authReducer = (state = initialState, action: TAction) => {
-
   switch (action.type) {
-
     case USER_LOGOUT: {
-      return initialState
+      return initialState;
     }
     case USER_DATA: {
       return {
         ...state,
         userData: action.payload,
         // isLoggedIn: true,
-      }
+      };
     }
     case LOADING_STATUS: {
       return {
         ...state,
         loading: action.payload,
-      }
+      };
     }
     default:
       return state;
